requirements.txt
wheel>=0.42.0
setuptools>=65.5.1
numpy>=2.2.0
streamlit>=1.32.0
pandas>=2.2.0
requests==2.31.0
opencv-python>=4.9.0
Pillow>=10.2.0
streamlit-extras>=0.3.6
streamlit-lottie>=0.0.5
scikit-learn>=1.4.0
joblib>=1.3.2
python-dotenv>=1.0.0
google-generativeai>=0.3.2
earthengine-api>=0.1.390
geemap>=0.29.6
folium>=0.14.0
streamlit-folium>=0.18.0

langchain_core
streamlit
langchain_community
pypdf
langchainhub
sentence_transformers
PyPDF2
langchain_ollama
pdfplumber
faiss-cpu
numpy
pandas
python-dotenv
yfinance
gtts 
speechrecognition 
pydub
